+++ src/backend/views/position.py
@@ -1,22 +1,20 @@
+# NEW CODE
+
 from datetime import datetime
 import uuid
 
 from flask import jsonify, request
 from flask_cognito import current_cognito_jwt
 from flask_smorest import Blueprint
-
 from backend.common.http import make_no_content_response
-from backend.common.exceptions import InvalidArgumentError
+from backend.common.exceptions import InvalidArgumentError, NotAllowedError
 from backend.models.db import transaction
 from backend.models.location import Location
 from backend.models.position import Position, PositionStatus, PositionRoleType, PositionSortFilter, PositionBookmark
 from backend.models.user import User
 from backend.views.base import AuthenticatedAPIBase
-
+from backend.common.permission import Permissions
 from backend.models.activities import Activity, ActivityGlobal, ActivityType
-
-"my code"
-from backend.models.notification import Notification, NotificationType
 
 
 blueprint = Blueprint('position', __name__, url_prefix='/positions')
@@ -25,7 +23,7 @@ blueprint = Blueprint('position', __name__, url_prefix='/positions')
 @blueprint.route('')
 class PositionAPI(AuthenticatedAPIBase):
     @staticmethod
-    def get():
+    def get():  # pylint: disable=duplicate-code
         sort = PositionSortFilter.lookup(request.args.get('sort')) if 'sort' in request.args else None
 
         keyword = request.args.get('keyword', None)
@@ -38,101 +36,94 @@ class PositionAPI(AuthenticatedAPIBase):
         with transaction() as tx:
             # This is the user making the request, for authorization purposes
             user = User.lookup(tx, current_cognito_jwt['sub'])
-            location = Location.lookup(tx, request.args.get('location_id')) if 'location_id' in request.args else None
+            current_user_role = Permissions.get_user_role(tx, current_cognito_jwt['sub'])
+            for permission in current_user_role.permissions:
+                if permission["api"] == 'Position' and permission['readPermission'] == 'Yes':
+                    location = Location.lookup(tx, request.args.get(
+                        'location_id')) if 'location_id' in request.args else None
 
-            positions = Position.search(
-                tx,
-                user,
-                sort=sort,
-                keyword=keyword,
-                department=department,
-                location=location,
-                role=role,
-                role_type=role_type,
-                pay_grade=pay_grade
-            )
-            positions = [position.as_dict() for position in positions]
-        return jsonify(positions)
+                    positions = Position.search(
+                        tx,
+                        user,
+                        sort=sort,
+                        keyword=keyword,
+                        department=department,
+                        location=location,
+                        role=role,
+                        role_type=role_type,
+                        pay_grade=pay_grade
+                    )
+                    positions = [position.as_dict() for position in positions]
+                    return jsonify(positions)
+            raise NotAllowedError(f"User '{user.id}' has no permission to see the position")
 
     @staticmethod
-    def post():
-        payload = request.json
-        # Generate a unique id for this position
-        position_id = str(uuid.uuid4())
-
-        now = datetime.utcnow()
-
-        required_fields = {
-            'manager_id', 'location_id', 'role_type', 'role', 'description'}
-        missing_fields = required_fields - set(payload.keys())
-        if missing_fields:
-            raise InvalidArgumentError(f"Field: {', '.join(missing_fields)} is required.")
-
-        roletype = PositionRoleType.validate_and_return_role_type(payload['role_type'])
-
+    def post():  # pylint: disable=too-many-locals
         with transaction() as tx:
-            hiring_manager = User.lookup(tx, payload['manager_id'])
-            location = Location.lookup(tx, payload['location_id'])
+            user = User.lookup(tx, current_cognito_jwt['sub'])
+            current_user_role = Permissions.get_user_role(tx, current_cognito_jwt['sub'])
+            for permission in current_user_role.permissions:
+                if permission["api"] == 'Position' and permission['createPermission'] == 'Yes':
+                    payload = request.json
+                    # Generate a unique id for this position
+                    position_id = str(uuid.uuid4())
 
-            pay_currency = payload['pay_currency'] if 'pay_currency' in payload else None
-            pay_minimum = payload['pay_minimum'] if 'pay_minimum' in payload else None
-            pay_maximum = payload['pay_maximum'] if 'pay_maximum' in payload else None
-            Position.validate_pay_range(pay_currency, pay_minimum, pay_maximum)
+                    now = datetime.utcnow()
 
-            position = Position(
-                id=position_id,
-                hiring_manager=hiring_manager,
-                role=payload.get('role'),
-                role_type=roletype,
-                department=payload['department'],
-                pay_currency=pay_currency,
-                pay_minimum=pay_minimum or 0,
-                pay_maximum=pay_maximum or 0,
-                location=location,
-                status=PositionStatus.ACTIVE.value,
-                created_at=now,
-                last_updated_at=now
-            )
-            position.update_details(payload)
-            tx.add(position)
-            #MY CODE:
-            #################
-            tx.add(Notification.add_notification(offer.offerer, NotificationType.NEW_APPLICATION, data={
-                'offer': {
-                    'offer_id': offer.id,
-                    'name': offer.name,
-                },
-                'proposal': {
-                    'proposal_id': proposal.id,
-                    'name': proposal.name,
-                },
-                'user': {
-                    'user_id': user.id,
-                    'name': user.name,
-                    'profile_picture_url': user.profile_picture_url
-                }
-            }))
-            send_slack_notification(offer.offerer, "Offer created successfully")
+                    required_fields = {
+                        'manager_id', 'location_id', 'role_type', 'role', 'description'}
+                    missing_fields = required_fields - set(payload.keys())
+                    if missing_fields:
+                        raise InvalidArgumentError(f"Field: {', '.join(missing_fields)} is required.")
 
+                    roletype = PositionRoleType.validate_and_return_role_type(payload['role_type'])
 
-            # Insert activity details in DB
-            activity_data = {
-                'position': {
-                    'offer_id': position.id
-                },
-                'user': {
-                    'user_id': position.hiring_manager.id,
-                    'name': position.hiring_manager.name,
-                    'profile_picture_url': position.hiring_manager.profile_picture_url
-                }
-            }
-            tx.add(Activity.add_activity(position.hiring_manager, ActivityType.NEW_POSITION_POSTED, data=activity_data))
-            tx.add(ActivityGlobal.add_activity(
-                position.hiring_manager.customer, ActivityType.NEW_POSITION_POSTED, data=activity_data))
+                    with transaction() as tx:
+                        hiring_manager = User.lookup(tx, payload['manager_id'])
+                        location = Location.lookup(tx, payload['location_id'])
 
-            position_details = position.as_dict()
+                        pay_currency = payload['pay_currency'] if 'pay_currency' in payload else None
+                        pay_minimum = payload['pay_minimum'] if 'pay_minimum' in payload else None
+                        pay_maximum = payload['pay_maximum'] if 'pay_maximum' in payload else None
+                        Position.validate_pay_range(pay_currency, pay_minimum, pay_maximum)
 
-        return position_details
+                        position = Position(
+                            id=position_id,
+                            hiring_manager=hiring_manager,
+                            role=payload.get('role'),
+                            role_type=roletype,
+                            department=payload['department'],
+                            pay_currency=pay_currency,
+                            pay_minimum=pay_minimum or 0,
+                            pay_maximum=pay_maximum or 0,
+                            location=location,
+                            status=PositionStatus.ACTIVE.value,
+                            created_at=now,
+                            last_updated_at=now
+                        )
+                        position.update_details(payload)
+                        tx.add(position)
+
+                        # Insert activity details in DB
+                        activity_data = {
+                            'position': {
+                                'offer_id': position.id
+                            },
+                            'user': {
+                                'user_id': position.hiring_manager.id,
+                                'name': position.hiring_manager.name,
+                                'profile_picture_url': position.hiring_manager.profile_picture_url
+                            }
+                        }
+                        tx.add(Activity.add_activity(position.hiring_manager, ActivityType.NEW_POSITION_POSTED,
+                                                     data=activity_data))
+                        tx.add(ActivityGlobal.add_activity(
+                            position.hiring_manager.customer, ActivityType.NEW_POSITION_POSTED, data=activity_data))
+
+                        position_details = position.as_dict()
+
+                    return position_details
+            raise NotAllowedError(f"User '{user.id}' has no permission for create position")
 
 
 @blueprint.route('/<position_id>')
@@ -140,66 +131,84 @@ class PositionByIdAPI(AuthenticatedAPIBase):
     @staticmethod
     def get(position_id):
         with transaction() as tx:
-            position = Position.lookup(tx, position_id)
-            return position.as_dict()
+            user = User.lookup(tx, current_cognito_jwt['sub'])
+            current_user_role = Permissions.get_user_role(tx, current_cognito_jwt['sub'])
+            for permission in current_user_role.permissions:
+                if permission["api"] == 'Position' and permission['readPermission'] == 'Yes':
+                    position = Position.lookup(tx, position_id, user)
+                    if Permissions.check_user_permission(position.hiring_manager.customer.id, user.customer_id):
+                        # print("---- positon customer id-----------",position.hiring_manager.customer.id)
+                        return position.as_dict()
+                    raise NotAllowedError("Invalid customer id")
+            raise NotAllowedError(f"User '{user.id}' has no permission to see the position id")
 
     @staticmethod
-    def put(position_id):
+    def put(position_id):  # noqa: C901
         now = datetime.utcnow()
 
         with transaction() as tx:
+            user = User.lookup(tx, current_cognito_jwt['sub'])
             position = Position.lookup(tx, position_id)
+            current_user_role = Permissions.get_user_role(tx, current_cognito_jwt['sub'])
+            for permission in current_user_role.permissions:
+                if permission["api"] == 'Position' and permission['createPermission'] == 'Yes':
+                    if Permissions.check_user_permission(position.hiring_manager.customer.id, user.customer_id):
+                        payload = request.json
 
-            payload = request.json
+                        if payload.get('manager_id'):
+                            position.hiring_manager = User.lookup(tx, payload['manager_id'])
 
-            if payload.get('manager_id'):
-                position.hiring_manager = User.lookup(tx, payload['manager_id'])
+                        if payload.get('role'):
+                            position.role = payload['role']
 
-            if payload.get('role'):
-                position.role = payload['role']
+                        if payload.get('role_type'):
+                            position.role_type = PositionRoleType.validate_and_return_role_type(payload['role_type'])
 
-            if payload.get('role_type'):
-                position.role_type = PositionRoleType.validate_and_return_role_type(payload['role_type'])
+                        if payload.get('department'):
+                            position.department = payload['department']
 
-            if payload.get('department'):
-                position.department = payload['department']
+                        Position.validate_pay_range(
+                            payload['pay_currency'] if 'pay_currency' in payload else position.pay_currency,
+                            payload['pay_minimum'] if 'pay_minimum' in payload else float(position.pay_minimum),
+                            payload['pay_maximum'] if 'pay_maximum' in payload else float(position.pay_maximum)
+                        )
+                        if 'pay_currency' in payload:
+                            position.pay_currency = payload['pay_currency']
+                        if 'pay_minimum' in payload:
+                            position.pay_minimum = payload['pay_minimum']
+                        if 'pay_maximum' in payload:
+                            position.pay_maximum = payload['pay_maximum']
 
-            Position.validate_pay_range(
-                payload['pay_currency'] if 'pay_currency' in payload else position.pay_currency,
-                payload['pay_minimum'] if 'pay_minimum' in payload else float(position.pay_minimum),
-                payload['pay_maximum'] if 'pay_maximum' in payload else float(position.pay_maximum)
-            )
-            if 'pay_currency' in payload:
-                position.pay_currency = payload['pay_currency']
-            if 'pay_minimum' in payload:
-                position.pay_minimum = payload['pay_minimum']
-            if 'pay_maximum' in payload:
-                position.pay_maximum = payload['pay_maximum']
-
-            if payload.get('location_id'):
-                position.location = Location.lookup(tx, payload['location_id'])
-
-            position.last_updated_at = now
-            position.update_details(payload)
-
-        # Fetch the position again to get updated response
-        with transaction() as tx:
-            position = Position.lookup(tx, position_id)
-            position_details = position.as_dict()
-        return position_details
+                        if payload.get('location_id'):
+                            position.location = Location.lookup(tx, payload['location_id'])
+                            position.last_updated_at = now
+                            position.update_details(payload)
+                        # Fetch the position again to get updated response
+                        with transaction() as tx:
+                            position = Position.lookup(tx, position_id)
+                            position_details = position.as_dict()
+                        return position_details
+                    raise NotAllowedError("Invalid customer id")
+            raise NotAllowedError(f"User '{user.id}' has no permission for update the position id")
 
     @staticmethod
     def delete(position_id):
         now = datetime.utcnow()
 
         with transaction() as tx:
+            user = User.lookup(tx, current_cognito_jwt['sub'])
             position = Position.lookup(tx, position_id)
+            current_user_role = Permissions.get_user_role(tx, current_cognito_jwt['sub'])
 
             # TODO: (santanu) System account can delete a position
-
-            position.status = PositionStatus.DELETED.value
-            position.last_updated_at = now
-        return make_no_content_response()
+            for permission in current_user_role.permissions:
+                if permission["api"] == 'Position' and permission['AdminPermission'] == 'Yes':
+                    if Permissions.check_user_permission(position.hiring_manager.customer.id, user.customer_id):
+                        position.status = PositionStatus.DELETED.value
+                        position.last_updated_at = now
+                        return make_no_content_response()
+                    raise NotAllowedError("Invalid customer id")
+            raise NotAllowedError(f"User '{user.id}' has no permission for delete the position id")
 
 
 @blueprint.route('/<position_id>/bookmark')
@@ -207,20 +216,30 @@ class PositionBookmarkAPI(AuthenticatedAPIBase):
     @staticmethod
     def put(position_id):
         with transaction() as tx:
-            position = Position.lookup(tx, position_id)
             user = User.lookup(tx, current_cognito_jwt['sub'])
-
-            bookmark = PositionBookmark.lookup(tx, user.id, position.id, must_exist=False)
-            if bookmark is None:
-                PositionBookmark(user=user, position=position, created_at=datetime.utcnow())
-        return make_no_content_response()
+            position = Position.lookup(tx, position_id)
+            current_user_role = Permissions.get_user_role(tx, current_cognito_jwt['sub'])
+            for permission in current_user_role.permissions:
+                if permission["api"] == 'Position' and permission['createPermission'] == 'No':
+                    if Permissions.check_user_permission(position.hiring_manager.customer.id, user.customer_id):
+                        bookmark = PositionBookmark.lookup(tx, user.id, position.id, must_exist=False)
+                        if bookmark is None:
+                            PositionBookmark(user=user, position=position, created_at=datetime.utcnow())
+                        return make_no_content_response()
+                    raise NotAllowedError("Invalid customer id")
+            raise NotAllowedError(f"User '{user.id}' has no permission for update the bookmark of position id")
 
     @staticmethod
     def delete(position_id):
         with transaction() as tx:
-            position = Position.lookup(tx, position_id)
             user = User.lookup(tx, current_cognito_jwt['sub'])
-
-            bookmark = PositionBookmark.lookup(tx, user.id, position.id)
-            tx.delete(bookmark)
-        return make_no_content_response()
+            position = Position.lookup(tx, position_id)
+            current_user_role = Permissions.get_user_role(tx, current_cognito_jwt['sub'])
+            for permission in current_user_role.permissions:
+                if permission["api"] == 'Position' and permission['AdminPermission'] == 'Yes':
+                    if Permissions.check_user_permission(position.hiring_manager.customer.id, user.customer_id):
+                        bookmark = PositionBookmark.lookup(tx, user.id, position.id)
+                        tx.delete(bookmark)
+                        return make_no_content_response()
+                    raise NotAllowedError("Invalid customer id")
+            raise NotAllowedError(f"User '{user.id}' has no permission for delete the position id")
